requirements.txt
django>=5.2,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-camel-case>=1.4
django-cors-headers>=4.3,<5.0
django-filter>=23.5
feedparser>=6.0
beautifulsoup4>=4.12
requests>=2.31
Pillow>=10.0
celery>=5.3
django-celery-beat>=2.5
apscheduler>=3.10
python-dotenv>=1.0
bleach>=6.1
httpx>=0.27
gunicorn>=21.2
